pyproject.toml
[tool.poetry]
packages = [{ include = "visualsplit", from = "." }]

[[tool.poetry.source]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"


[tool.poetry.dependencies]
torch = {source = "pytorch-cu118"}
torchvision = {source = "pytorch-cu118"}
[project]
name = "visualsplit"
version = "0.1.0"
description = ""
authors = [
    { name = "HenryQUQ", email = "quchenyuan@qq.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch (==2.2.1+cu118)",
    "torchvision (==0.17.1+cu118)",
    "transformers (==4.38.2)",
    "accelerate (==0.28.0)",
    "kornia (==0.7.1)",
    "pytorch-ignite (==0.4.13)",
    "wandb (==0.16.4)",
    "datasets (==3.1.0)",
    "opencv-python (==4.9.0.80)",
    "numpy (>=1.26.3,<1.27.0)",
    "pillow (>=10.2.0,<10.3.0)",
    "matplotlib (>=3.8.3,<3.9.0)",
    "requests (>=2.32.3,<2.33.0)",
    "tqdm (>=4.67.0,<4.68.0)",
    "pyyaml (>=6.0.1,<6.1.0)",
    "packaging (>=24.0,<25.0)",
    "lpips (>=0.1.4,<0.2.0)",
    "safetensors (>=0.4.2,<0.5.0)",
    "bitsandbytes (>=0.44.1,<0.45.0)",
    "notebook (>=7.4.5,<8.0.0)",
    "diffusers (>=0.35.0,<0.36.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "black (>=25.1.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
